+++ src/feelpp/benchmarking/reframe/setup.py
@@ -1,9 +1,9 @@
 
-from feelpp.benchmarking.feelpp_toolboxes.reframe.parameters import NbTasks
-from feelpp.benchmarking.feelpp_toolboxes.config.configReader import ConfigReader
-from feelpp.benchmarking.feelpp_toolboxes.config.configSchemas import ConfigFile,MachineConfig
-from feelpp.benchmarking.feelpp_toolboxes.reframe.validation import ValidationHandler
-from feelpp.benchmarking.feelpp_toolboxes.reframe.scalability import ScalabilityHandler
+from feelpp.benchmarking.reframe.parameters import NbTasks
+from feelpp.benchmarking.reframe.config.configReader import ConfigReader
+from feelpp.benchmarking.reframe.config.configSchemas import ConfigFile,MachineConfig
+from feelpp.benchmarking.reframe.validation import ValidationHandler
+from feelpp.benchmarking.reframe.scalability import ScalabilityHandler
 
 import reframe as rfm
 import os
