+++ src/feelpp/benchmarking/report/templates/benchmark.adoc.j2
@@ -32,7 +32,7 @@
 
 [%dynamic%close,python]
 ----
-from feelpp.benchmarking.feelpp_toolboxes.reframe.report import Report
+from feelpp.benchmarking.reframe.report import Report
 report=Report(file_path="{{ filepath }}")
 ----
